splitting Auth interface from Cognito implementation
