+++ src/control-plane/index.ts
@@ -1,7 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './auth';
+export * from './auth/auth';
+export * from './auth/cognito-auth';
 export * from './control-plane-api';
 export * from './control-plane';
 export * from './lambda-layers';
